+++ aura/lib/Aura/Diff.hs
@@ -1,6 +1,6 @@
 -- |
 -- Module    : Aura.Diff
--- Copyright : (c) Colin Woodbury, 2012 - 2019
+-- Copyright : (c) Colin Woodbury, 2012 - 2020
 -- License   : GPL3
 -- Maintainer: Colin Woodbury <colin@fosskers.ca>
 --