+++ aura/lib/Aura/Pkgbuild/Editing.hs
@@ -4,7 +4,7 @@
 
 -- |
 -- Module    : Aura.Pkgbuild.Base
--- Copyright : (c) Colin Woodbury, 2012 - 2019
+-- Copyright : (c) Colin Woodbury, 2012 - 2020
 -- License   : GPL3
 -- Maintainer: Colin Woodbury <colin@fosskers.ca>
 --