+++ aura/lib/Aura/Utils.hs
@@ -2,7 +2,7 @@
 
 -- |
 -- Module    : Aura.Utils
--- Copyright : (c) Colin Woodbury, 2012 - 2019
+-- Copyright : (c) Colin Woodbury, 2012 - 2020
 -- License   : GPL3
 -- Maintainer: Colin Woodbury <colin@fosskers.ca>
 --
@@ -203,7 +203,7 @@ yesNoPrompt ss msg = do
 
 -- | An empty response emplies "yes".
 isAffirmative :: Language -> Text -> Bool
-isAffirmative l t = T.null t || elem t (yesPattern l)
+isAffirmative l t = T.null t || elem (T.toCaseFold t) (yesPattern l)
 
 -- | Doesn't prompt when `--noconfirm` is used.
 optionalPrompt :: Settings -> (Language -> Doc AnsiStyle) -> IO Bool